# api/openapi.yaml
openapi: 3.1.0
info:
  title: DinnerOS API
  version: 0.5.0
  description: |
    Backend for the DinnerOS iOS app. "DinnerOS" is a working name.

    Conventions (see docs/api.md):
    - All product routes are versioned under `/api/v1`.
    - JSON bodies use camelCase. Unknown request fields are rejected.
    - Every non-2xx response uses the `ErrorResponse` envelope.
    - Authenticated routes take `Authorization: Bearer <accessToken>`
      (see docs/authentication.md).
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/households/{householdId}/pantry:
    parameters:
      - $ref: "#/components/parameters/HouseholdID"
    get:
      summary: List pantry items
      description: |
        Every matching item in the household's pantry, in aisle order and then
        by name. A pantry holds at most 1000 items, so the list isn't
        paginated. Requires `household.view`.
      operationId: listPantryItems
      tags: [pantry]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/PantryStatus"
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/GroceryCategoryName"
        - name: staple
          in: query
          description: "`true` for staples only, `false` for the rest."
          schema:
            type: boolean
        - name: q
          in: query
          description: |
            Text the display name contains (case-insensitive), or that the
            normalized name contains (`jalapeno` finds "Jalapeño"). Literal
            text, not a pattern.
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: The pantry items.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PantryItemListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      summary: Add a pantry item
      description: |
        Adds an ingredient by catalog `ingredientId` or free-text `name`, or
        merges into the item that already has the same ingredient (one item per
        ingredient per household). Requires `pantry.edit`.

        A free-text name that is a catalog ingredient is linked to it. A merge
        applies `status` (default `in_stock`) and every field sent; omitted
        fields keep their values and the display name is kept, so retrying is
        safe.
      operationId: addPantryItem
      tags: [pantry]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPantryItemRequest"
      responses:
        "201":
          description: Added a new item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PantryItem"
        "200":
          description: Merged into the existing item for this ingredient.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PantryItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/PantryConflict"
  /api/v1/households/{householdId}/pantry/{itemId}:
    parameters:
      - $ref: "#/components/parameters/HouseholdID"
      - $ref: "#/components/parameters/PantryItemID"
    patch:
      summary: Update a pantry item
      description: |
        Changes the fields sent. `quantity`, `expiresOn`, and `note` may be
        `null` (or `""`) to clear them; clearing `quantity` also clears `unit`.
        Setting `status` to `out` clears the amount. Requires `pantry.edit`.
      operationId: updatePantryItem
      tags: [pantry]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePantryItemRequest"
      responses:
        "200":
          description: The updated item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PantryItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/PantryConflict"
    delete:
      summary: Delete a pantry item
      description: Requires `pantry.edit`.
      operationId: deletePantryItem
      tags: [pantry]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/households/{householdId}/pantry/bulk:
    parameters:
      - $ref: "#/components/parameters/HouseholdID"
    post:
      summary: Set several items' statuses
      description: |
        Sets up to 200 items' statuses at once, for example after shopping.
        Items marked `out` lose their amount. Requested IDs that aren't in the
        pantry are listed in `missing` instead of failing the request.
        Requires `pantry.edit`.
      operationId: setPantryStatuses
      tags: [pantry]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkPantryStatusRequest"
      responses:
        "200":
          description: The updated items, in request order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkPantryStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/households/{householdId}/pantry/staples/defaults:
    parameters:
      - $ref: "#/components/parameters/HouseholdID"
    post:
      summary: Add the default staples
      description: |
        Adds salt, black pepper, cooking oil, olive oil, butter, sugar, flour,
        garlic powder, and onion powder, in stock and marked `isStaple`, unless
        the pantry already has them. Each links to the catalog ingredient with
        its name or a known alias ("Pepper" for black pepper). Idempotent:
        staples already in the pantry under either name are left unchanged and
        counted in `skipped`. No request body. Requires `pantry.edit`.
      operationId: addDefaultPantryStaples
      tags: [pantry]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The staples added.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultStaplesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        DinnerOS access token from a sign-in or refresh response. Valid for 15
        minutes. An expired token yields 401 `token_expired`; refresh and retry.
...
  parameters:
    HouseholdID:
      name: householdId
      in: path
      required: true
      schema:
        type: string
        example: 66e5a1f2c3b4a5d6e7f80913
...
    PantryItemID:
      name: itemId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    GroceryCategoryName:
      type: string
      enum: [produce, meat-seafood, dairy-eggs, bakery, deli, pantry, spices, condiments, frozen, beverages, other]
    PantryStatus:
      type: string
      enum: [in_stock, low, out]
      description: |
        `in_stock`: grocery lists show the ingredient as `inPantry`. `low`:
        not treated as at home (`pantryHint` if every recipe marks it as a
        staple, otherwise `toBuy`). `out`: always `toBuy`.
    PantryItem:
      type: object
      required: [id, householdId, ingredientId, key, displayName, category, quantity, quantityValue, unit, status, isStaple, expiresOn, note, updatedBy, createdAt, updatedAt]
      properties:
        id:
          type: string
        householdId:
          type: string
        ingredientId:
          type: [string, "null"]
          description: The catalog ingredient; null for free text that matched no catalog ingredient.
        key:
          type: string
          description: The normalized ingredient name, unique within the household (the catalog key when linked).
          example: olive oil
        displayName:
          type: string
          example: Olive Oil
        category:
          $ref: "#/components/schemas/GroceryCategoryName"
        quantity:
          type: [string, "null"]
          description: Exact amount as "n" or "n/d"; null when no amount was recorded ("have some").
          example: "3/2"
        quantityValue:
          type: [number, "null"]
          description: The same amount as a number, for display only.
          example: 1.5
        unit:
          type: [string, "null"]
          description: DinnerOS unit code; null exactly when quantity is null.
          example: cup
        status:
          $ref: "#/components/schemas/PantryStatus"
        isStaple:
          type: boolean
          description: An always-have item such as salt or oil.
        expiresOn:
          type: [string, "null"]
          format: date
        note:
          type: string
        updatedBy:
          type: string
          description: User ID of the last change.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PantryItemListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PantryItem"
    AddPantryItemRequest:
      type: object
      description: Either `ingredientId` or `name` is required.
      properties:
        ingredientId:
          type: string
          description: A catalog ingredient ID from GET /api/v1/ingredients.
        name:
          type: string
          maxLength: 100
          description: Display name. Defaults to the catalog name when `ingredientId` is given.
        category:
          $ref: "#/components/schemas/GroceryCategoryName"
        quantity:
          type: string
          maxLength: 32
          description: A positive amount such as `2`, `0.5`, `1/2`, `1 1/2`, or `½`. Omit for "have some".
          example: 1 1/2
        unit:
          type: string
          description: DinnerOS unit code (docs/grocery-engine.md#units). Defaults to `count` when `quantity` is given; requires `quantity`.
          example: cup
        status:
          $ref: "#/components/schemas/PantryStatus"
        isStaple:
          type: boolean
        expiresOn:
          type: string
          format: date
        note:
          type: string
          maxLength: 500
    UpdatePantryItemRequest:
      type: object
      minProperties: 1
      properties:
        displayName:
          type: string
          maxLength: 100
        category:
          $ref: "#/components/schemas/GroceryCategoryName"
        quantity:
          type: [string, "null"]
          description: A positive amount; null or "" clears the amount and unit.
        unit:
          type: [string, "null"]
          description: DinnerOS unit code for the current or new quantity.
        status:
          $ref: "#/components/schemas/PantryStatus"
        isStaple:
          type: boolean
        expiresOn:
          type: [string, "null"]
          format: date
        note:
          type: [string, "null"]
          maxLength: 500
    BulkPantryStatusRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 200
          items:
            type: object
            required: [id, status]
            properties:
              id:
                type: string
              status:
                $ref: "#/components/schemas/PantryStatus"
    BulkPantryStatusResponse:
      type: object
      required: [items, missing]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PantryItem"
        missing:
          type: array
          description: Requested IDs that aren't in the household's pantry.
          items:
            type: string
    DefaultStaplesResponse:
      type: object
      required: [items, skipped]
      properties:
        items:
          type: array
          description: The staples added.
          items:
            $ref: "#/components/schemas/PantryItem"
        skipped:
          type: integer
          description: Default staples the pantry already had.
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
              example: not_found
            message:
              type: string
              description: Human-readable description. Not stable; do not parse.
              example: resource not found
            requestId:
              type: string
              description: Matches the X-Request-ID response header; quote it when reporting problems.
              example: 5f0c1e9a2b7d4c3e8f6a1b2c
  responses:
    Forbidden:
      description: "`forbidden`: the caller is a member of the household, but their role doesn't allow this action."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    PantryConflict:
      description: "`conflict`: the item kept changing concurrently after several automatic retries; retry."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: The resource does not exist or is not visible to the caller.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: |
        `invalid_request` (malformed JSON, unknown field, wrong type, empty body)
        or `validation_failed` (missing or invalid field).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthenticated:
      description: |
        `unauthenticated` (missing, invalid, rotated, or revoked token) or
        `token_expired` (access token expired; refresh and retry).
      headers:
        WWW-Authenticate:
          description: Sent by authenticated routes.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...